refactor detection pipeline
