+++ camera_functions.py
@@ -14,21 +14,19 @@ class Camera:
         else:
             self.camera = cv2.VideoCapture(0)
 
-    def capture_rasberrypi(self, save_path, delay_s):
+    def capture_and_save_rasberrypi(self, save_path, delay_s):
         self.camera.start_preview()
-        sleep(delay_s)
         self.camera.capture(save_path)
         self.camera.stop_preview()
 
-    def capture_windows(self, save_path, delay_s):
+    def capture_and_save_windows(self, save_path, delay_s):
         if os.path.isfile(save_path):
             os.remove(save_path)
         retval, frame = self.camera.read()
         cv2.imwrite(save_path, frame)
-        sleep(delay_s)
 
-    def capture_with_delay(self, save_path, delay_s):
+    def capture_and_save(self, save_path, delay_s):
         if self.isRasberryPI:
-            self.capture_rasberrypi(save_path, delay_s)
+            self.capture_and_save_rasberrypi(save_path, delay_s)
         else:
-            self.capture_windows(save_path, delay_s)
+            self.capture_and_save_windows(save_path, delay_s)
