+++ network_cam.py
@@ -1,10 +1,10 @@
 import requests
 import json
+from constants import const
 
 
-def post_data(people_count, device_type, zone_id):
+async def post_data(people_count, device_type, zone_id):
     data = {"deviceId": device_type, "placeId": zone_id, "count": people_count}
     data = json.dumps(data)
-    addr = "http://192.168.1.64:8080/recreation/api/data"
     headers = {'content-type': 'application/json'}
-    r = requests.post(url=addr, headers=headers, data=data)
+    requests.post(url=const.recreation_server_endpoint, headers=headers, data=data)
